feature/ATC-53 - Adds aircraftConstants and updates refs to use new file.
