+++ src/assets/scripts/airport/Arrival/ArrivalBase.js
@@ -17,7 +17,7 @@ import {
     calculateDistanceToBoundary,
     calculateHeadingFromTwoPositions
 } from '../../math/flightMath';
-import { FLIGHT_CATEGORY } from '../../aircraft/AircraftInstanceModel';
+import { FLIGHT_CATEGORY } from '../../constants/aircraftConstants';
 import { AIRPORT_CONSTANTS } from '../../constants/airportConstants';
 import { TIME } from '../../constants/globalConstants';
 import { LOG } from '../../constants/logLevel';
